Move `OrderBySelect` into own component
